rename components and tweak save theme button styling
