+++ src/app/(theming)/(ecommerce-sample)/page.tsx
@@ -1,4 +1,4 @@
-import SampleUI from "@/components/sample-ui/sample-ui";
+import SampleUI from "@/components/sample-ui/sample-ecommerce-page";
 
 export default function EcommerceSampleUi() {
   return (
