+++ src/components/sample-ui/sample-hero-section.tsx
