+++ src/components/theme-options/save-theme.tsx
@@ -1,7 +1,7 @@
 export default function SaveTheme() {
   return (
-    <footer className="sticky bottom-0 mt-12 flex justify-end bg-white/60 px-8 pt-2 pb-6 backdrop-blur-sm">
-      <button className="inline-flex w-full items-center justify-center whitespace-nowrap rounded-full bg-neutral-800 px-4 py-2 font-medium text-white transition-colors hover:bg-neutral-950 focus-visible:outline-none focus-visible:ring-1 disabled:pointer-events-none disabled:opacity-50 ">
+    <footer className="sticky bottom-0 mt-12 flex justify-end bg-white/60 px-8 pb-6 pt-2 backdrop-blur-sm">
+      <button className="inline-flex w-full items-center justify-center whitespace-nowrap rounded-full border border-neutral-200 bg-neutral-50 shadow-sm px-4 py-3 font-medium text-lg text-neutral-600 transition-colors hover:bg-neutral-800 hover:text-neutral-50 focus-visible:outline-none focus-visible:ring-1 disabled:pointer-events-none disabled:opacity-50">
         Save theme
       </button>
     </footer>
